user_service/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
idna==3.10
py-zipkin==1.2.8
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
